feat(deprecated_topic): gate v8 topic compat behind a build tag
Official binaries still ship deprecated-topic support (goreleaser,
CI test matrix and the CLAUDE.md build-tag list all include
`deprecated_topic`), but a library-user `go build` with no tags now
produces a hub that does NOT carry the v8 matcher shim:

  - matcher_deprecated.go (deprecatedMatcher + resolveDeprecatedStringClaim)
    and subscribematchers_deprecated.go (appendDeprecatedTopicMatchers)
    are `//go:build deprecated_topic`.
  - matcher_notdeprecated.go / subscribematchers_notdeprecated.go carry
    `//go:build !deprecated_topic` stubs that refuse bare-string JWT
    claims and the `topic=` query parameter regardless of
    WithProtocolVersionCompatibility.
  - matcher.go's resolveMatcherClaims and subscribematchers.go's
    parseMatchers call those hooks instead of referencing
    deprecatedMatcher directly.

Tests are reorganised so the default `go test` still compiles:

  - Dedicated-v8 tests moved to `_deprecated_test.go` files
    (subscriber_deprecated_test.go, subscription_deprecated_test.go,
    hubmatcher_deprecated_test.go, subscribematchers_deprecated_test.go,
    matcher_claims_deprecated_test.go) and tagged deprecated_topic.
  - `createDummy` no longer forces WithProtocolVersionCompatibility(8)
    (modern by default); a new `createDeprecatedDummy` helper in
    hub_deprecated_topic_test.go provides the v8 variant for tagged
    tests, alongside `createDeprecatedAuthorizedJWT` (v8 string
    claims vs `createDummyAuthorizedJWT`'s modern Exact claims).
  - `subscribeRecorder` moved to its own untagged helper file so hub
    tests can keep using it independently of subscribe_test.go.
  - `subscribe_test.go` is tagged deprecated_topic — every test there
    exercises the `?topic=` query parameter and therefore only makes
    sense under compat mode.
  - `TestMatchTopic`, `TestSubscriptionHandler`,
    `TestSubscriptionsHandlerForTopic`, `TestParseMatchersDeprecatedTopic`
    and the v8 default-matchers / route-gating tests now live in the
    tagged files; `caddy/caddy_deprecated_test.go` also grew a build
    tag so it's skipped in the default build.

Wiring:

  - .github/workflows/ci.yml, .goreleaser.yml and CLAUDE.md all have
    `deprecated_topic` added to the default tag list so current CI /
    official binaries keep v8 compat; dropping the tag in a future
    release is the way to kill the code path for good.
