+++ .github/workflows/ci.yml
@@ -13,7 +13,7 @@ permissions:
 
 env:
   GOTOOLCHAIN: local
-  GOFLAGS: "-tags=deprecated_server,deprecated_transport,nobadger,nomysql,nopgx"
+  GOFLAGS: "-tags=deprecated_server,deprecated_transport,deprecated_topic,nobadger,nomysql,nopgx"
 
 jobs:
   golangci:
