+++ .goreleaser.yml
@@ -21,6 +21,7 @@ builds:
       - -X 'github.com/caddyserver/caddy/v2.CustomVersion=Mercure.rocks {{ .Version }} Caddy'
     tags:
       - deprecated_transport
+      - deprecated_topic
       - nobadger
       - nomysql
       - nopgx
@@ -47,6 +48,7 @@ builds:
       - -s -w -X github.com/dunglas/mercure/common.version={{ .Version }} -X github.com/dunglas/mercure/common.commit={{ .ShortCommit }} -X github.com/dunglas/mercure/common.buildDate={{ .Date }}
     tags:
       - deprecated_transport
+      - deprecated_topic
       - deprecated_server
     goos:
       - linux
