+++ CLAUDE.md
@@ -3,10 +3,10 @@
 ## Build Tags
 
 Always pass the following build tags when building or testing unless explicitly asked not to:
-`deprecated_server,deprecated_transport,nobadger,nomysql,nopgx`
+`deprecated_server,deprecated_transport,deprecated_topic,nobadger,nomysql,nopgx`
 
 For example:
 
 ```console
-go test -tags "deprecated_server,deprecated_transport,nobadger,nomysql,nopgx" ./...
+go test -tags "deprecated_server,deprecated_transport,deprecated_topic,nobadger,nomysql,nopgx" ./...
 ```
