+++ local_bench_test.go
@@ -48,9 +48,9 @@ func subBenchLocalTransport(b *testing.B, topics, concurrency, matchPct int, tes
 	for i := range concurrency {
 		s := NewLocalSubscriber("", slog.Default(), tss)
 		if i%100 < matchPct {
-			s.setMatchers(stringsToDeprecatedMatchers(tsMatch), stringsToDeprecatedMatchers(nil))
+			s.setMatchers(stringsToExactMatchers(tsMatch), stringsToExactMatchers(nil))
 		} else {
-			s.setMatchers(stringsToDeprecatedMatchers(tsNoMatch), stringsToDeprecatedMatchers(nil))
+			s.setMatchers(stringsToExactMatchers(tsNoMatch), stringsToExactMatchers(nil))
 		}
 
 		subscribers[i] = s
