+++ publish_test.go
@@ -23,8 +23,8 @@ func TestPublish(t *testing.T) {
 
 		topics := []string{"https://example.com/books/1"}
 		s := NewLocalSubscriber("", slog.Default(), &TopicSelectorStore{})
-		s.setMatchers(stringsToDeprecatedMatchers(topics), stringsToDeprecatedMatchers(topics))
-		s.Claims = &claims{Mercure: mercureClaim{Subscribe: stringsToDeprecatedClaims(topics)}}
+		s.setMatchers(stringsToExactMatchers(topics), stringsToExactMatchers(topics))
+		s.Claims = &claims{Mercure: mercureClaim{Subscribe: stringsToExactClaims(topics)}}
 
 		require.NoError(t, hub.transport.AddSubscriber(t.Context(), s))
 
@@ -266,8 +266,8 @@ func TestPublishHandlerOK(t *testing.T) {
 
 		topics := []string{"https://example.com/books/1"}
 		s := NewLocalSubscriber("", slog.Default(), &TopicSelectorStore{})
-		s.setMatchers(stringsToDeprecatedMatchers(topics), stringsToDeprecatedMatchers(topics))
-		s.Claims = &claims{Mercure: mercureClaim{Subscribe: stringsToDeprecatedClaims(topics)}}
+		s.setMatchers(stringsToExactMatchers(topics), stringsToExactMatchers(topics))
+		s.Claims = &claims{Mercure: mercureClaim{Subscribe: stringsToExactClaims(topics)}}
 
 		require.NoError(t, hub.transport.AddSubscriber(t.Context(), s))
 
@@ -342,7 +342,7 @@ func TestPublishHandlerGenerateUUID(t *testing.T) {
 		topics := []string{"https://example.com/books/1"}
 
 		s := NewLocalSubscriber("", slog.Default(), &TopicSelectorStore{})
-		s.setMatchers(stringsToDeprecatedMatchers(topics), stringsToDeprecatedMatchers(topics))
+		s.setMatchers(stringsToExactMatchers(topics), stringsToExactMatchers(topics))
 
 		require.NoError(t, h.transport.AddSubscriber(t.Context(), s))
