+++ subscribematchers_deprecated.go
@@ -1,15 +1,11 @@
+//go:build deprecated_topic
+
 package mercure
 
-import "errors"
-
-// errTopicParamNotSupported is returned when the deprecated "topic" query
-// parameter is used without WithProtocolVersionCompatibility.
-var errTopicParamNotSupported = errors.New(`the "topic" query parameter is not supported anymore, use "match" instead`)
-
 // appendDeprecatedTopicMatchers handles the v8 "topic" query parameter by
-// wrapping each value in a deprecatedMatcher-backed topicMatcher. Only
-// reachable when backward compatibility with protocol v8 is enabled; see
-// parseMatchers.
+// wrapping each value in a deprecatedMatcher-backed topicMatcher. Reachable
+// from parseMatchers only when WithProtocolVersionCompatibility(8) is set
+// AND the hub was compiled with the deprecated_topic build tag.
 func (h *Hub) appendDeprecatedTopicMatchers(matchers []topicMatcher, values []string) ([]topicMatcher, error) {
 	for _, v := range values {
 		if len(v) > maxPatternLength {
