+++ subscriber_bench_test.go
@@ -93,7 +93,7 @@ func subBenchSubscriber(b *testing.B, topics, concurrency, matchPct int, testNam
 		}
 	}
 
-	s.setMatchers(stringsToDeprecatedMatchers(ts), stringsToDeprecatedMatchers(nil))
+	s.setMatchers(stringsToExactMatchers(ts), stringsToExactMatchers(nil))
 	b.Cleanup(s.Disconnect)
 
 	ctx, done := context.WithCancel(ctx)
