+++ subscriberecorder_test.go
@@ -1,0 +1,56 @@
+package mercure
+
+import (
+	"net/http/httptest"
+	"os"
+	"time"
+)
+
+// subscribeRecorder is a ResponseRecorder that also implements SetWriteDeadline
+// / Flush — used by tests that drive SubscribeHandler against a synthetic
+// response writer.
+type subscribeRecorder struct {
+	*httptest.ResponseRecorder
+
+	writeDeadline time.Time
+}
+
+func newSubscribeRecorder() *subscribeRecorder {
+	return &subscribeRecorder{ResponseRecorder: httptest.NewRecorder()}
+}
+
+func (r *subscribeRecorder) SetWriteDeadline(deadline time.Time) error {
+	if deadline.After(r.writeDeadline) {
+		r.writeDeadline = deadline
+	}
+
+	return nil
+}
+
+func (r *subscribeRecorder) Write(buf []byte) (int, error) {
+	if time.Now().After(r.writeDeadline) {
+		return 0, os.ErrDeadlineExceeded
+	}
+
+	return r.ResponseRecorder.Write(buf)
+}
+
+func (r *subscribeRecorder) WriteString(str string) (int, error) {
+	if time.Now().After(r.writeDeadline) {
+		return 0, os.ErrDeadlineExceeded
+	}
+
+	return r.WriteString(str)
+}
+
+func (r *subscribeRecorder) FlushError() error {
+	if time.Now().After(r.writeDeadline) {
+		return os.ErrDeadlineExceeded
+	}
+
+	r.Flush()
+
+	return nil
+}
+
+func (*subscribeRecorder) Flush() {}
